requirements.txt
Pillow
torch
torchaudio
torchvision
cmake
yacs
albumentations
gdown
omegaconf
pandas
nibabel
pytorch-lightning
PyYAML
scikit-image
scikit-learn
scipy
sklearn
protobuf
timm
wandb
matplotlib
ipywidgets
kornia
einops
segmentation-models-pytorch
nipype
SimpleITK
antspyx